Fixed dotfiles bootstrapping

+++ script/bootstrap.sh
@@ -1,8 +1,6 @@
 #!/usr/bin/env bash
 
 function bootstrap_dotfiles() {
-	local exclude_paths dotfile target link link_directory_path
-
 	cd $HOME/.dotfiles
 
 	if [[ $(nmcli --colors no networking connectivity) == 'full' ]]; then
@@ -13,25 +11,32 @@ function bootstrap_dotfiles() {
 		return 1
 	fi
 
-    exclude_paths=$(echo $HOME/.dotfiles/{.git,install,script})
-
-	for dotfile in $(find $HOME/.dotfiles -type f -not -exec realpath --relative-base $HOME/.dotfiles {} \; $(printf -- " -or -path %s -prune" $exclude_paths)); do
+    local dotfile target target_directory link link_directory
+	for dotfile in $(find $HOME/.dotfiles -type f -exec realpath --relative-base $HOME/.dotfiles {} \;); do
         target="$HOME/.dotfiles/$dotfile"
-		link="$HOME/$dotfile"
-		if [[ ! -h $link ]]; then
-			link_directory_path=$(dirname $link)
-			if [[ ! -d $link_directory_path ]]; then
-				if [[ -e $link ]]; then
-					rm $link_directory_path
-				fi
-				mkdir -p $link_directory_path
-			fi
+        target_directory=$(dirname $target)
+        if [[ ! (\
+            $target_directory =~ ^$HOME/.dotfiles/.git || \
+            $target_directory =~ ^$HOME/.dotfiles/install || \
+            $target_directory =~ ^$HOME/.dotfiles/script || \
+            $target == $HOME/.dotfiles/.gitignore \
+        ) ]]; then
+            link="$HOME/$dotfile"
+            if [[ ! -h $link ]]; then
+                link_directory=$(dirname $link)
+                if [[ ! -d $link_directory ]]; then
+                    if [[ -e $link ]]; then
+                        rm $link_directory
+                    fi
+                    mkdir -p $link_directory
+                fi
 
-			if [[ -e $link ]]; then
-				rm $link
-			fi
-			ln -s $target $link
-		fi
+                if [[ -e $link ]]; then
+                    rm $link
+                fi
+                ln -s $target $link
+            fi
+        fi
 	done
 
 	cd "$OLDPWD"
